+++ bot.py
@@ -9,7 +9,7 @@ from helpers.db_helper import DBHelper
 from helpers.gdrive_helper import GDriveHelper
 from datetime import datetime
 from discord.errors import NotFound
-from games.blackjack import Blackjack
+from games import blackjack
 
 class NotBotAdmin(commands.CheckFailure):
     pass
@@ -127,6 +127,8 @@ class Melbot():
                 await self.bot.process_commands(message)
 
         # --- bot commands ---
+        blackjack.add_bot_commands(self.bot, self.playing_blackjack, self.db)
+
         self.bot.remove_command('help')
         @self.bot.command(help="Display the help message.")
         async def help(ctx):
@@ -237,85 +239,7 @@ class Melbot():
                 await ctx.send(f"{ctx.author} - You lost {points} points.")
             else:
                 await ctx.send(f"{ctx.author} - You won {earned_points} points.")
-
-        @self.bot.command(help="Play a game of blackjack for melpoints. Syntax: !blackjack <melpoints>")
-        async def blackjack(ctx, points: int = None):
-            user_id = ctx.author.id
-            blackjack = Blackjack(user_id)
-            user_points = await self.db.get_total_currency(str(user_id))
-
-            # Validate points
-            if points is None or type(points) != int:
-                await ctx.send("Please provide a number of melpoints to bet. Syntax: !blackjack <melpoints>")
-                return
-            if points > user_points:
-                await ctx.send(f"You do not have enough melpoints to bet {points} points. You have {user_points} melpoints.")
-                return
-            if points < blackjack.config['min_bet']:
-                await ctx.send(f"The minimum bet is {blackjack.config['min_bet']} points.")
-                return
-            if points > blackjack.config['max_bet']:
-                await ctx.send(f"The maximum bet is {blackjack.config['max_bet']} points.")
-                return
-
-            if user_id in self.playing_blackjack:
-                await ctx.send("You are already playing a game of blackjack.")
-                return
-
-            blackjack.deal()
-
-            await ctx.send(f"""{ctx.author.name} - you drew: {blackjack.players[user_id].hand[0]} and {blackjack.players[user_id].hand[1]}\nI drew: {blackjack.players["dealer"].hand[0]} and something else.\n\nYou have {blackjack.calculate_score(user_id)} points. Do you want to !hit or !stand?""")
-            self.playing_blackjack.update({user_id: {"bet": points, "game": blackjack, "ctx": ctx,"start_time": datetime.now().timestamp()}})
-            
-        @self.bot.command()
-        async def hit(ctx):
-            user_id = ctx.author.id
-            if user_id not in self.playing_blackjack:
-                await ctx.send("You are not playing blackjack.")
-                return
-            blackjack = self.playing_blackjack[user_id]["game"]
-            blackjack.hit(user_id)
-            score = blackjack.calculate_score(user_id)
-            if score > 21:
-                await ctx.send(f"{ctx.author.name} - you drew: {blackjack.players[user_id].hand[-1]}\n\nYou have {score} points. You busted!")
-                await self.db.add_event(user_id, self.playing_blackjack[user_id]["bet"] * -1, 'blackjack')
-                self.playing_blackjack.pop(user_id)
-                return
-            await ctx.send(f"{ctx.author.name} - you drew: {blackjack.players[user_id].hand[-1]}\n\nYou have {score} points. Do you want to !hit or !stand?")
-            self.playing_blackjack[user_id].update({"game": blackjack})
-
-        @self.bot.command()
-        async def stand(ctx):
-            user_id = ctx.author.id
-            if user_id not in self.playing_blackjack:
-                await ctx.send("You are not playing blackjack.")
-                return
-            blackjack = self.playing_blackjack[user_id]["game"]
-            user_score = blackjack.calculate_score(user_id)
-            dealer_score = blackjack.calculate_score("dealer")
-            while dealer_score < 17:
-                blackjack.hit("dealer")
-                await ctx.send(f"The dealer drew: {blackjack.players['dealer'].hand[-1]}")
-                dealer_score = blackjack.calculate_score("dealer")
-                await asyncio.sleep(0.5)
-            if dealer_score > 21:
-                await ctx.send(f"{ctx.author.name} - you have {user_score} points. The dealer busted with {dealer_score} points. You win!")
-                await self.db.add_event(user_id, self.playing_blackjack[user_id]["bet"] * (blackjack.config["payout"] - 1), 'blackjack')
-            elif user_score > dealer_score:
-                await ctx.send(f"{ctx.author.name} - you have {user_score} points. The dealer has {dealer_score} points. You win!")
-                await self.db.add_event(user_id, self.playing_blackjack[user_id]["bet"] * (blackjack.config["payout"] - 1), 'blackjack')
-            elif user_score < dealer_score:
-                await ctx.send(f"{ctx.author.name} - you have {user_score} points. The dealer has {dealer_score} points. You lose!")
-                await self.db.add_event(user_id, self.playing_blackjack[user_id]["bet"] * -1, 'blackjack')
-            elif user_score == 21 and len(blackjack.players[user_id].hand) == 2 and user_score > dealer_score:
-                await ctx.send(f"{ctx.author.name} - you have {user_score} points. You got a blackjack! You win!")
-                await self.db.add_event(user_id, self.playing_blackjack[user_id]["bet"] * (blackjack.config["payout"] - 1), 'blackjack')
-            else:
-                await ctx.send(f"{ctx.author.name} - you have {user_score} points. The dealer has {dealer_score} points. It's a tie! But the house always wins.")
-                await self.db.add_event(user_id, self.playing_blackjack[user_id]["bet"] * -1, 'blackjack')
-            self.playing_blackjack.pop(user_id)
  
-
         @self.bot.command(help="Display the leaderboard.")
         async def leaderboard(ctx):
             leaderboard = await self.db.get_leaderboard()
